docker-flask/requirements.txt
Flask==0.12
Jinja2==2.11
mysql-connector-python==8.0.19